# feat: Optimize em.hydrate performance