--- a/packages/codegen/src/symbols.ts
+++ b/packages/codegen/src/symbols.ts
@@ -60,6 +60,7 @@ export const newChangesProxy = imp("newChangesProxy@joist-orm");
 export const LoadHint = imp("LoadHint@joist-orm");
 export const Loaded = imp("Loaded@joist-orm");
 export const isLoaded = imp("isLoaded@joist-orm");
+export const getOrmField = imp("getOrmField@joist-orm");
 export const isEntity = imp("isEntity@joist-orm");
 export const getEm = imp("getEm@joist-orm");
 export const loadLens = imp("loadLens@joist-orm");