This PR changes entity construction to just wrap the incoming database row/tuple, and only populate the internal `__orm.data` map lazily/as individual fields are accessed.

It also renames the `__orm` field that was hidden (`enumerable: false`) via an expensive `Object.defineProperty` in the `BaseEntity` constructor to a `#orm` private field that's exposed by a `getOrmField` static field + function.

This takes loading 50k books from ~250ms -> 180ms, a 30% increase.

See the `hydrate-perf.mkd` for each of the incremental changes/improvements.

We're still quite a bit slower than just reading raw records off the wire, but it's better.

Two observations:

1. Dates are expensive, and
2. Relations are expensive

Per dates, the next hot-spot is actually date parsing inside of pg-types/postgres-date, which in theory is a cost that both knex and em.find are paying. Supposedly this has been fixed if we just bump to the latest pg-types:

https://transporeon-visibility-hub.medium.com/improving-date-parsing-in-node-postgres-by-4x-fa3506b4400c

https://github.com/bendrucker/postgres-date/pull/21

Per relations, notices how loading 50k authors is 525ms but loading 50k books is 250ms, nearly 50% cheaper. I'm fairly sure the flamegraph shows this is due to initialization the many boundary-case-testing relations in `Author.ts` that we know we want to make lazy.

My only hesitation is that postgres.js itself is ~2x faster loading 50k books than 50k authors, so it's also probably just the fact that books is a small table & authors is not (more columns to test more boundary conditions).

---

Baseline on main:

```
benchmark        time (avg)             (min … max)       p75       p99      p995
--------------------------------------------------- -----------------------------
• loading 50k authors
--------------------------------------------------- -----------------------------
postgres.js  163.05 ms/iter  (150.9 ms … 178.34 ms) 173.76 ms 178.34 ms 178.34 ms
knex         141.42 ms/iter (135.47 ms … 154.87 ms) 142.52 ms 154.87 ms 154.87 ms
em.find      520.96 ms/iter  (464.55 ms … 633.1 ms) 526.74 ms  633.1 ms  633.1 ms

summary for loading 50k authors
  knex
   1.15x faster than postgres.js
   3.68x faster than em.find

• loading 50k books
--------------------------------------------------- -----------------------------
postgres.js   67.47 ms/iter   (58.4 ms … 120.25 ms)  65.96 ms 120.25 ms 120.25 ms
knex         110.88 ms/iter  (109.02 ms … 113.4 ms) 111.82 ms  113.4 ms  113.4 ms
em.find      248.46 ms/iter (243.45 ms … 256.17 ms) 251.01 ms 256.17 ms 256.17 ms

summary for loading 50k books
  postgres.js
   1.64x faster than knex
   3.68x faster than em.find
```

---

With this branch:

```
benchmark        time (avg)             (min … max)       p75       p99      p995
--------------------------------------------------- -----------------------------
• loading 50k authors
--------------------------------------------------- -----------------------------
postgres.js  167.68 ms/iter (148.23 ms … 214.48 ms) 173.35 ms 214.48 ms 214.48 ms
knex         149.73 ms/iter (140.71 ms … 164.54 ms)  158.2 ms 164.54 ms 164.54 ms
em.find      364.53 ms/iter (303.51 ms … 534.84 ms) 363.99 ms 534.84 ms 534.84 ms

summary for loading 50k authors
  knex
   1.12x faster than postgres.js
   3.43x faster than em.find

• loading 50k books
--------------------------------------------------- -----------------------------
postgres.js   63.32 ms/iter   (59.97 ms … 70.43 ms)  64.83 ms  70.43 ms  70.43 ms
knex         128.38 ms/iter (111.54 ms … 227.91 ms)  126.4 ms 227.91 ms 227.91 ms
em.find      186.05 ms/iter (178.08 ms … 207.55 ms) 189.58 ms 207.55 ms 207.55 ms

summary for loading 50k books
  postgres.js
   2.03x faster than knex
   2.94x faster than em.find
```