@stephenh 2023-12-26
:tada: This PR is included in version 1.136.0 :tada:

The release is available on:
- `v1.136.0`
- [GitHub release](https://github.com/stephenh/joist-ts/releases/tag/v1.136.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: